feat: 初始化 ts 重构
