+++ .prettierrc.js
@@ -1,0 +1,12 @@
+module.exports = {
+  printWidth: 100,
+  semi: true,
+  singleQuote: true,
+  trailingComma: 'all',
+  bracketSpacing: false,
+  jsxBracketSameLine: true,
+  arrowParens: 'avoid',
+  insertPragma: false,
+  tabWidth: 2,
+  useTabs: false,
+};
