+++ package.json
@@ -1,42 +1,50 @@
 {
-    "name": "matman",
-    "private": true,
-    "scripts": {
-        "publish": "git pull && lerna publish --cd-version patch",
-        "force-publish": "lerna publish --cd-version patch --force-publish=*",
-        "docs:dev": "vuepress dev docs",
-        "docs:build": "vuepress build docs",
-        "docs:deploy": "npm run docs:build && gh-pages -d docs-dist",
-        "test": "node ./test/run-test.js",
-        "test:direct": "node ./node_modules/mocha/bin/_mocha",
-        "test:direct:debugnightmare": "cross-env DEBUG=nightmare npm run test:direct"
-    },
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/matmanjs/matman.git"
-    },
-    "keywords": [
-        "e2e",
-        "test"
-    ],
-    "author": "helinjiang",
-    "license": "MIT",
-    "bugs": {
-        "url": "https://github.com/matmanjs/matman/issues"
-    },
-    "dependencies": {
-        "element-ui": "^2.13.0",
-        "fs-extra": "^9.0.0",
-        "less": "^3.11.1",
-        "less-loader": "^5.0.0",
-        "vue-particles": "^1.0.9"
-    },
-    "devDependencies": {
-        "chai": "^4.2.0",
-        "cross-env": "^7.0.2",
-        "gh-pages": "^2.2.0",
-        "lerna": "^3.20.2",
-        "mocha": "^7.1.2",
-        "vuepress": "^1.3.1"
-    }
+  "name": "matman",
+  "private": true,
+  "scripts": {
+    "init": "rm -rf node_modules && npm install && npm run clean && npm run bootstrap",
+    "bootstrap": "lerna bootstrap",
+    "clean": "lerna clean --yes",
+    "publish": "git pull && lerna publish --cd-version patch",
+    "force-publish": "lerna publish --cd-version patch --force-publish=*",
+    "docs:dev": "vuepress dev docs",
+    "docs:build": "vuepress build docs",
+    "docs:deploy": "npm run docs:build && gh-pages -d docs-dist",
+    "test": "node ./test/run-test.js",
+    "test:direct": "node ./node_modules/mocha/bin/_mocha",
+    "test:direct:debugnightmare": "cross-env DEBUG=nightmare npm run test:direct"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/matmanjs/matman.git"
+  },
+  "keywords": [
+    "e2e",
+    "test"
+  ],
+  "author": "helinjiang",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/matmanjs/matman/issues"
+  },
+  "dependencies": {
+    "element-ui": "^2.13.0",
+    "fs-extra": "^9.0.0",
+    "less": "^3.11.1",
+    "less-loader": "^5.0.0",
+    "vue-particles": "^1.0.9"
+  },
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^3.1.0",
+    "@typescript-eslint/parser": "^3.1.0",
+    "chai": "^4.2.0",
+    "cross-env": "^7.0.2",
+    "eslint": "^7.2.0",
+    "gh-pages": "^2.2.0",
+    "lerna": "^3.20.2",
+    "mocha": "^7.1.2",
+    "prettier": "^2.0.5",
+    "typescript": "^3.9.5",
+    "vuepress": "^1.3.1"
+  }
 }
