--- packages/matman/.babelrc
@@ -1,15 +1,0 @@
-{
-  "presets": [
-    "env",
-    "stage-0"
-  ],
-  "plugins": [
-    [
-      "transform-runtime",
-      {
-        "polyfill": false,
-        "regenerator": true
-      }
-    ]
-  ]
-}
