+++ packages/matman/package.json
@@ -27,22 +27,14 @@
   },
   "homepage": "https://github.com/matmanjs/matman#readme",
   "dependencies": {
-    "babel-runtime": "^6.26.0",
     "fs-extra": "^9.0.0",
     "glob": "^7.1.6",
-    "istanbul": "^0.4.5",
-    "istanbul-lib-coverage": "^3.0.0",
-    "istanbul-lib-source-maps": "^4.0.0",
     "lodash": "^4.17.15",
     "matman-crawler": "^5.0.7",
     "nightmare-handler": "^1.0.10"
   },
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.3",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-stage-0": "^6.24.1",
+    "@types/lodash": "^4.14.155",
     "chai": "^4.2.0",
     "mocha": "^7.1.2"
   }
